fix bg zoom and extract out getMediaOrigin util
