+++ assets/src/animation/effects/backgroundZoom/index.js
@@ -19,7 +19,7 @@
  */
 import { BG_MIN_SCALE, BG_MAX_SCALE, SCALE_DIRECTION } from '../../constants';
 import { AnimationZoom } from '../../parts/zoom';
-import { lerp } from '../../utils';
+import { lerp, getMediaOrigin } from '../../utils';
 
 export function EffectBackgroundZoom({
   element,
@@ -33,6 +33,9 @@ export function EffectBackgroundZoom({
   // at element scale 100, the range should be [1, 4]
   const range = [BG_MIN_SCALE / element.scale, BG_MAX_SCALE / element.scale];
 
+  // Account for moving bg media relative to frame
+  const origin = getMediaOrigin(element);
+
   return AnimationZoom({
     zoomFrom: lerp(zoomDirection === SCALE_DIRECTION.SCALE_OUT ? 1 : 0, range),
     zoomTo: 1,
@@ -40,5 +43,6 @@ export function EffectBackgroundZoom({
     delay,
     easing,
     targetLeafElement: true,
+    transformOrigin: `${origin.horizontal}% ${origin.vertical}%`,
   });
 }
