+++ assets/src/animation/utils/index.js
@@ -20,3 +20,4 @@ export { getTotalDuration } from './getTotalDuration';
 export { getMediaBoundOffsets, hasOffsets } from './mediaPositions';
 export { getExclusion, orderByKeys } from './objectOperations';
 export { clamp, lerp, progress } from './range';
+export { getMediaOrigin } from './getMediaOrigin';
